rfid-reader/requirements.txt
RPi.GPIO==0.6.3
Requests>=2.21.0
